requirements.txt
torchvision
torch
numpy
pytz
scipy
smart_open
Pillow
scikit_learn
